Disable the select patchsets dialog
Github doesn't have a notion of patchsets so this is not going to work
out of the box. Still, we may want to keep this feature around if we
ever get to implement something similar in hubtty so comment it out for
now until we have a better idea what to do with it.