Break dependency between commit and messages
Following the pattern we used for comments in fb6c709f, let's do the
same for messages (PR reviews and comments, as opposed to
inline-comments).  This allows to save all messages to the database now
that they're no longer tied to a commit anymore, since commits could
disappear in the event of a rebase.