Display all inline comments in change view
This adapts the view to show all inline comments in the change view.
Previously, only the comments for the commits currently belonging to the
PR were shown. This means we would lose comments in case of a rebase.

Also fix `AttributeError: 'NoneType' object has no attribute 'sha'`
error that was possible due to message no longer always having an
associated commit object.