Best guess which inline comment goes with which review
Since github does not link inline comments to code review, we use the
creation timestamp to make the connection. However, review and inline
comments often end up with different timestamps wo we have to guess
which message belongs to which code review. We currently assume that
inline messages within 5 seconds of the review belong to the review.