Add reference to commit in approval
This allows to know if the approval is for the most recent PR or not.

And because an approval might have been submitted on a commit that no
longer is part of the PR, we have to save the commit sha rather than
the commit record id since there is no guarantee this have been saved
to the DB, for instance in case of a rebase.