Fix query when subscribing to more than one repo
When passing more than one query to the `query()` method, it was wrongly
returning the entire json instead of only the items. The response looks
like:

  {
    "total_count": 7,
    "incomplete_results": false,
    "items": [
      ...
    ]
  }

This caused a TypeError when iterating over the results.

Also the method was wrongly joining the queries and was causing github
to return unexpected results.

Instead, send a search query for each subscribed repository.