Initial support for rate limiting
Naive implementation of rate limiting support where we check the
`X-RateLimit-Remaining` header for each query and set hubtty offline
when the remaining requests in the current limit window get below the
arbitrarily chosen 10 value.

Possible improvements would be to use the value of `X-RateLimit-Reset`
to schedule next query instead of hardcoding the value to 30 seconds.

Also, we should make use of conditional requests to lower the number of
queries [1].

[1] https://docs.github.com/en/rest/overview/resources-in-the-rest-api#conditional-requests