Better handle rate limiting
The Search API has a rate limit of 30 requests per minute, which
differs from the normal Core API where the limit is 5000 requests per
hour.

We now requeue the request when hitting rate limit and retry at the
time indicated by the `X-RateLimit-Reset` header.

This needs to be ported to other requests than GET.