Filter our invalid changes from query results
As the number of projects one subscribes to grows, we're more likely to
hit github rate limit, which may end up with incomplete entries in the
database.

This commit ensures that Changes are valid and syncs the change
otherwise.

Another better approach would be to treat Change sync as an atomic
operation.

Fixes #15