Remove approval labels
We're not using the label and permitted_label tables. There were used in
gertty for approval labels. Remove these tables as they make things
confusing with github labels, which are tags we can attach to PRs.

This commit also restore the "Code-Review" summary from the change
lists.

We'll bring back the label table once we properly implement github
labels.