Do not assume PR has commits
It may be possible that a PR was rebased and no longer contain any
commits. When synchronizing changes, we need to stop assuming PR
contains commits. We don't need the same protection while displaying
changes since github automatically closes PRs without commits and as
a consequence never make it to the change view.

Fixes #24