Do not assume commits are unique
We wrongly assume only one commit with the same SHA existed in the
commit table and failed to synchronize changes when it contained commits
from another PR.

With this change it's now possible to get commits with the same sha in
the commit table.
As a consequence, move the DatabaseSession.getCommitBySha() method to
Change.getCommitBySha() and rename getCommitBySha() to getCommitsBySha()
in DatabaseSession.

Closes #26