Discard non-available repositories
It may be that the repositories added from the config file with
additional-repositories may be private or do not exist. In this case
github returns a 404 and hubtty now discards these repositories.

This is made possible thanks to the introduction of a custom
`response_callback` to the Sync.get() method.

Closes #30