Fix Cancel button for merge dialog
The cancel button (or the previous screen keybinding for that matter)
was returning to the project list due to previous screen not part of
the screen stack when executing the callback. We need to use a lambda
to execute backScreen in the right context.

We need to apply the same change to other dialogues in this view and
eventually port the same fix to gertty.