Drop Sync Query
This is used to sync changes for repos we're not subscribed to. In
gertty it syncs changes we're author of, and starred changes.

We're not using it at the moment in hubtty, and not planning to: it's
very likely that hubtty users have submitted changes to many repos and
this feature is going to cause hubtty to clone many, potentially big,
repositories.

We can re-evaluate if there's demand for this feature.