Drop cherry-pick
Github does not yet offer an easy way to cherry-pick a PR to
a different branch [1] and implementing the feature myself is more work
than I'm willing to do at the moment while we're getting towards
a first release. So let's just remove this code for now and revisit
later.

[1] https://github.community/t/do-a-cherry-pick-via-the-api/14573/4