Drop SyncChangeByNumberTask
Again, this was used in gertty to sync changes in repos the user hasn't
subscribed to, as a result of a query, for instance if a commit message
referenced a change in another repo. This seems like a nice idea,
however this can quickly become messy with ETOOMANYCLONEDREPOS.

We may consider bringing this feature back if there's demand.