Fix pagination for search queries
Github API returns a dict for search queries, unlike other queries that
lead to pagination. This lead to us only returning the last page of
result in case we got a paginated search result.

To fix the issue, we build a list with all results - stored under the
items key for search results - and return that.

Closes #61