Drop cherry-pick from pr list view
It doesn't make much sense to allow cherry-picking the most recent
commit of a PR if we don't get to see the list of commits the PR
contains. If users want to cherry-pick the most recent commit locally,
they can still do it from the pull request view. They can also
cherry-pick any of the commits contained in the PR locally.