Fix retrieving draft message
Creating a new comments would wrongly create multiple draft messages.

Draft message is the message that is currently being worked on, while
pending message is the same message, but finalized (user submitted the
message) waiting to be uploaded.

We're not using commit.getPendingMessage() other than for checking if
the message is still in draft or ready to be uploaded, and we might
want to rename it hasPendingMessage() to clarify.