Rename commit.getPendingMessage() to hasPendingMessage()
We were only using the commit.getPendingMessage() method to know if
there were pending message attached to the commit, and not actually
looking at the content of the pending message.

To make it clearer, let's rename the method to hasPendingMessage().  If
we want get the content of the message to be uploaded, we can use
getDraftMessage(), as message remains in draft until it is uploaded.