Pre-select existing approval in review dialog
Fix the value of the current approval state to match what github
expects as a valid event. Typically, COMMENTED becomes COMMENT.

This in turn allows the dialog to pre-select existing approval.

Also fix an issue if approval status is empty, or if there is another
approval for a different commit of the same PR.