Post comments for all commits
Previously, only comments for the last commit were pushed to github. This means user could lose comments.

This changes the logic and create one review per commit that has
comments as the API doesn't expose a way to push comments to different
commits in the same review. It must be possible however, as the Github
UI allows posting comments to different commit in the same review.