Let user know about token issue on 403
Github may return a 403 error when using a token generated via the
device flow, and the organization has enabled 3rd party application
restriction. In which case we should use a personal access token.

With this commit we now inform the user about the token issue and
redirect him to the relevant hubtty documentation. The request will no
longer be retried until hubtty restarts.

Fixes #20.