Highlight draft PRs
Based on gertty's commits 94310529ed and 363114eb7a.

Sadly, the github REST API does not yet allow changing the draft status
of the PRs [1] so we're limiting ourselves to showing draft PRs. We
would have to use their GraphQL API [2] for changing the draft status of
PRs, similar to what the github CLI client does [3].

Fixes #41.

[1] https://docs.github.com/en/rest/reference/pulls#update-a-pull-request
[2] https://docs.github.com/en/graphql
[3] https://github.com/cli/cli/blob/c8d5a6b/api/queries_pr.go#L687-L704