Register palette entry on new label
When hubtty detects a label it doesn't know about, it should create the
corresponding palette entry.

Sadly, it appears urwid does not offer the ability to update or delete
a palette entry. We can't update the color of an existing label when it
was changed in github. We'll have to wait for the user to restart hubtty
to get the new color.